refactor(node/state_delta): extract store_setup submodule (increment 4)
Continues the state_delta god-file split. Moves the delta-store setup cluster
into `store_setup.rs` — behavior-preserving, pure code movement
(mod.rs: +4 wiring lines, -128).

Moved:
- DeltaStoreSetup (struct; fields `pub(super)` for the apply-path destructure)
- choose_owned_identity (also reached by the sibling `buffering` module via
  `super::`, re-exported through mod.rs's import)
- init_delta_store

Explicit imports (no `use super::*`), per the established convention. Call
sites, the sibling buffering module, and the in-module tests are unchanged.

Validated: node lib tests (304) and sync_sim (314) pass.

Remaining: the two orchestrators (handle_state_delta,
apply_authorized_state_delta) stay in mod.rs by design; request_missing_deltas
and ensure_application_available remain (sync/blob concerns).
