+++ crates/node/src/handlers/state_delta/mod.rs
@@ -22,6 +22,7 @@ use crate::utils::choose_stream;
 mod buffering;
 mod crypto;
 mod events;
+mod store_setup;
 mod verify;
 
 pub(crate) use buffering::{
@@ -37,6 +38,9 @@ use events::{
     emit_state_mutation_event_parsed, execute_cascaded_events, execute_event_handlers_parsed,
     parse_events_payload, CascadeOutcome,
 };
+// `choose_owned_identity` is also reached by the sibling `buffering` module via
+// `super::` (re-exported through this import).
+use store_setup::{choose_owned_identity, init_delta_store, DeltaStoreSetup};
 pub(crate) use verify::{verify_position_group_id_matches_context, GroupIdCheck};
 
 pub(crate) struct StateDeltaMessage {
@@ -1073,134 +1077,6 @@ pub async fn handle_state_delta(
         false,
     )
     .await
-}
-
-struct DeltaStoreSetup {
-    store: DeltaStore,
-    is_uninitialized: bool,
-}
-
-async fn choose_owned_identity(
-    context_client: &ContextClient,
-    context_id: &ContextId,
-) -> Result<PublicKey> {
-    let identities = context_client.get_context_members(context_id, Some(true));
-    let Some((our_identity, _)) = choose_stream(identities, &mut rand::thread_rng())
-        .await
-        .transpose()?
-    else {
-        bail!("no owned identities found for context: {}", context_id);
-    };
-
-    Ok(our_identity)
-}
-
-async fn init_delta_store(
-    node_state: &crate::NodeState,
-    node_clients: &crate::NodeClients,
-    context_id: ContextId,
-    our_identity: PublicKey,
-    root_hash: Hash,
-    sync_timeout: std::time::Duration,
-) -> Result<DeltaStoreSetup> {
-    let is_uninitialized = root_hash == Hash::default();
-
-    let (delta_store_ref, is_new_store) = {
-        let mut is_new = false;
-        let delta_store = node_state
-            .delta_stores
-            .entry(context_id)
-            .or_insert_with(|| {
-                is_new = true;
-                DeltaStore::new(
-                    [0u8; 32],
-                    node_clients.context.clone(),
-                    context_id,
-                    our_identity,
-                )
-            });
-
-        (delta_store.clone(), is_new)
-    };
-
-    if is_new_store {
-        let init_result = async {
-            // `load_persisted_deltas` surfaces any records with
-            // `applied: true, events: Some(..)` — crash-leftovers
-            // whose handlers never completed. Merged with the normal
-            // cascade events below so a single handler pass covers both
-            // (#2185). Share the DB scan with the DAG restore to avoid
-            // a second full-table iteration (#2194 review).
-            let pending_handler_events = match delta_store_ref.load_persisted_deltas().await {
-                Ok(result) => {
-                    if !result.pending_handler_events.is_empty() {
-                        info!(
-                            %context_id,
-                            pending_count = result.pending_handler_events.len(),
-                            "Replaying handlers interrupted by crash before events were cleared"
-                        );
-                    }
-                    result.pending_handler_events
-                }
-                Err(e) => {
-                    warn!(
-                        ?e,
-                        %context_id,
-                        "Failed to load persisted deltas, starting with empty DAG"
-                    );
-                    Vec::new()
-                }
-            };
-
-            let missing_result = delta_store_ref.get_missing_parents().await;
-            if !missing_result.missing_ids.is_empty() {
-                warn!(
-                    %context_id,
-                    missing_count = missing_result.missing_ids.len(),
-                    "Missing parents after loading persisted deltas - will request from network"
-                );
-            }
-
-            // The two sources are disjoint by construction:
-            // `pending_handler_events` are records that were `applied:
-            // true` on disk before this init ran, so they're restored
-            // into the DAG as already-applied by `load_persisted_deltas`
-            // and can't show up in `get_missing_parents`'s
-            // pending→applied diff. Concat directly.
-            let mut events_to_run = missing_result.cascaded_events;
-            events_to_run.extend(pending_handler_events);
-
-            execute_cascaded_events(
-                &events_to_run,
-                &node_clients.node,
-                &node_clients.context,
-                &context_id,
-                &our_identity,
-                sync_timeout,
-                "initial load",
-                None,
-                &delta_store_ref,
-            )
-            .await
-        }
-        .await;
-
-        if let Err(err) = init_result {
-            warn!(
-                %context_id,
-                ?err,
-                "Initial delta store setup failed - removing store to retry on next delta"
-            );
-            // Remove the store so the next delta triggers a fresh init with retry
-            node_state.delta_stores.remove(&context_id);
-            return Err(err);
-        }
-    }
-
-    Ok(DeltaStoreSetup {
-        store: delta_store_ref,
-        is_uninitialized,
-    })
 }
 
 #[cfg(test)]
